+++ CHANGELOG.md
@@ -4,9 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.4.0] - 2021-09-09
 ### Added
 - `liblogc_config` for configuration parsing using libconfig
+- log function with critical level `log_fatal` and `fatal` macro that calls
+  `exit()` to terminate execution of application.
 - log\_argp argument `--no-syslog` for daemon logs as there would be no way to
   disable default syslog otherwise.